Lecture 68. Spring MVC Controller Advice - added MvcExceptionHandler and updated the existing controller
