+++ src/main/java/guru/springframework/sbmbrewery/web/controller/CustomerController.java
@@ -54,15 +54,4 @@ public class CustomerController {
     public void deleteCustomer(@PathVariable("customerId") UUID customerId) {
         customerService.deleteById(customerId);
     }
-
-    @ExceptionHandler({ConstraintViolationException.class})
-    public ResponseEntity<List<String>> validationErrorHandler(ConstraintViolationException e) {
-        List<String> errors = new ArrayList<>(e.getConstraintViolations().size());
-
-        e.getConstraintViolations()
-            .forEach(constraintViolation ->
-                     errors.add(constraintViolation.getPropertyPath() + ":" + constraintViolation.getMessage()));
-
-        return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
-    }
 }
