+++ src/main/java/guru/springframework/sbmbrewery/web/controller/MvcExceptionHandler.java
@@ -1,0 +1,33 @@
+package guru.springframework.sbmbrewery.web.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
+import org.springframework.validation.ObjectError;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import javax.validation.ConstraintViolationException;
+import java.util.ArrayList;
+import java.util.List;
+
+@ControllerAdvice
+public class MvcExceptionHandler {
+
+    @ExceptionHandler(ConstraintViolationException.class)
+    public ResponseEntity<List<String>> validationErrorHandler(ConstraintViolationException e) {
+        List<String> errors = new ArrayList<>(e.getConstraintViolations().size());
+
+        e.getConstraintViolations()
+                .forEach(constraintViolation ->
+                        errors.add(constraintViolation.getPropertyPath() + ":" + constraintViolation.getMessage()));
+
+        return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(BindException.class)
+    public ResponseEntity<List<ObjectError>> handleBindException(BindException e) {
+        return new ResponseEntity<>(e.getAllErrors(), HttpStatus.BAD_REQUEST);
+    }
+
+}
